Create LocalFileHeaderAccessor and enhance ZipFile class
Introduced a new utility class named LocalFileHeaderAccessor to aid in the interaction with Local File Header (LFH) data. This class provides a way to retrieve various properties from the LFH. At the same time, additional methods were added to the ZipFile class for providing access to individual file metadata and its content as a string.